refactor: split four 1000+ LoC files via part-of extensions
Flatten the long tail in lib/. All four splits are behavior-preserving:
classes / extensions move into part files in subfolders, public API and
private-field access stay identical via `part of`.

- import_service.dart 1788 → 986: asset-event flow moved to a part
  (extension AssetImportFlow on ImportService).
- asset_detail_screen.dart 1324 → 365: chart section / composition
  section / edit dialog → asset_detail/*.dart.
- assets_screen.dart 1078 → 233: _AssetTile + _CreateAssetDialog →
  assets/*.dart.
- account_detail_screen.dart 1289 → 1046: entry types + transfer tile
  → entries.dart; period header + totals chip → list_widgets.dart.

Distribution across lib/:
  files > 1000 LoC: 9 → 6
  max:              1788 → 1493
  p90:              974  → 850
  avg LoC/file:     354  → 329

All four suites green: analyze clean, 778 unit, integration +4, live
data fetch +1.
