+++ lib/ui/screens/account_detail_screen.dart
@@ -20,6 +20,9 @@ import '../widgets/selection/selectable_item.dart';
 import '../widgets/selection/selection_action_bar.dart';
 import '../widgets/selection/selection_controller.dart';
 
+part 'account_detail/entries.dart';
+part 'account_detail/list_widgets.dart';
+
 final _log = getLogger('AccountDetailScreen');
 
 /// Sentinel account id used by the virtual "All accounts" entry. The detail
@@ -1040,250 +1043,4 @@ class _AccountDetailScreenState extends ConsumerState<AccountDetailScreen> {
 
 /// A row in the All-accounts list. Either a single transaction or an
 /// inter-account transfer that collapses two opposing legs into one row.
-sealed class _Entry {
-  DateTime get valueDate;
-}
 
-class _TxEntry extends _Entry {
-  final Transaction tx;
-  _TxEntry(this.tx);
-  @override
-  DateTime get valueDate => tx.valueDate;
-}
-
-class _TransferEntry extends _Entry {
-  final Transaction inflow; // amount > 0
-  final Transaction outflow; // amount < 0
-  _TransferEntry({required this.inflow, required this.outflow});
-  @override
-  DateTime get valueDate => outflow.valueDate;
-  String get currency => inflow.currency;
-  double get absAmount => inflow.amount.abs();
-}
-
-class _TransferTile extends StatefulWidget {
-  final _TransferEntry entry;
-  final Map<int, String> accountNameById;
-  final String locale;
-  final AppStrings s;
-  /// Builds the row for a single leg of the transfer using the same widget as
-  /// regular transactions, so the expanded view is visually identical to the
-  /// rest of the list.
-  final Widget Function(Transaction) legTileBuilder;
-  const _TransferTile({
-    required this.entry,
-    required this.accountNameById,
-    required this.locale,
-    required this.s,
-    required this.legTileBuilder,
-  });
-
-  @override
-  State<_TransferTile> createState() => _TransferTileState();
-}
-
-class _TransferTileState extends State<_TransferTile> {
-  bool _expanded = false;
-
-  @override
-  Widget build(BuildContext context) {
-    final scheme = Theme.of(context).colorScheme;
-    final entry = widget.entry;
-    final s = widget.s;
-    final amtFmt = fmt.currencyFormat(widget.locale, entry.currency);
-    final dateFmt = fmt.shortDateFormat(widget.locale);
-    final fromName = widget.accountNameById[entry.outflow.accountId] ??
-        '#${entry.outflow.accountId}';
-    final toName = widget.accountNameById[entry.inflow.accountId] ??
-        '#${entry.inflow.accountId}';
-    final blue = scheme.primary;
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        ListTile(
-          dense: true,
-          leading: CircleAvatar(
-            radius: 16,
-            backgroundColor: blue.withValues(alpha: 0.12),
-            child: Icon(Icons.swap_horiz, size: 16, color: blue),
-          ),
-          title: Text(
-            s.transferLabel,
-            maxLines: 1,
-            overflow: TextOverflow.ellipsis,
-            style: const TextStyle(fontSize: 14, fontWeight: FontWeight.w600),
-          ),
-          subtitle: Row(
-            children: [
-              Text(dateFmt.format(entry.valueDate),
-                  style: const TextStyle(fontSize: 12)),
-              const SizedBox(width: 6),
-              Flexible(
-                child: Text(
-                  s.transferFromTo(fromName, toName),
-                  style: TextStyle(
-                    fontSize: 12,
-                    color: scheme.onSurfaceVariant,
-                    fontStyle: FontStyle.italic,
-                  ),
-                  maxLines: 1,
-                  overflow: TextOverflow.ellipsis,
-                ),
-              ),
-            ],
-          ),
-          trailing: Row(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              PrivacyText(
-                amtFmt.format(entry.absAmount),
-                style: TextStyle(
-                  fontWeight: FontWeight.bold,
-                  color: blue,
-                  fontSize: 14,
-                ),
-              ),
-              Icon(
-                _expanded ? Icons.expand_less : Icons.expand_more,
-                size: 18,
-                color: scheme.onSurfaceVariant,
-              ),
-            ],
-          ),
-          onTap: () => setState(() => _expanded = !_expanded),
-        ),
-        AnimatedSize(
-          duration: const Duration(milliseconds: 180),
-          curve: Curves.easeInOut,
-          alignment: Alignment.topCenter,
-          child: _expanded
-              ? Padding(
-                  padding: const EdgeInsets.only(left: 24),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.stretch,
-                    children: [
-                      widget.legTileBuilder(entry.outflow),
-                      const Divider(height: 1),
-                      widget.legTileBuilder(entry.inflow),
-                    ],
-                  ),
-                )
-              : const SizedBox.shrink(),
-        ),
-      ],
-    );
-  }
-}
-
-class _PeriodHeader extends StatelessWidget {
-  final String label;
-  final Map<String, ({double income, double expense})> totals;
-  final String locale;
-  final bool isMonth;
-  const _PeriodHeader({
-    required this.label,
-    required this.totals,
-    required this.locale,
-    required this.isMonth,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final scheme = Theme.of(context).colorScheme;
-    // Sort currencies for stable display; primary first by total absolute flow.
-    final entries = totals.entries.toList()
-      ..sort((a, b) {
-        final fa = a.value.income - a.value.expense; // expense is negative
-        final fb = b.value.income - b.value.expense;
-        return fb.abs().compareTo(fa.abs());
-      });
-    return Container(
-      width: double.infinity,
-      padding: EdgeInsets.symmetric(
-        horizontal: 16,
-        vertical: isMonth ? 8 : 6,
-      ),
-      color: isMonth
-          ? scheme.surfaceContainerHigh
-          : scheme.surfaceContainerHighest,
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          Expanded(
-            child: Text(
-              label,
-              style: Theme.of(context).textTheme.labelSmall?.copyWith(
-                    color: scheme.onSurfaceVariant,
-                    fontWeight: isMonth ? FontWeight.w700 : FontWeight.w600,
-                    fontSize: isMonth ? 13 : null,
-                  ),
-            ),
-          ),
-          if (entries.isNotEmpty)
-            Column(
-              crossAxisAlignment: CrossAxisAlignment.end,
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                for (final e in entries)
-                  _TotalsChip(
-                    income: e.value.income,
-                    expense: e.value.expense,
-                    currency: e.key,
-                    locale: locale,
-                  ),
-              ],
-            ),
-        ],
-      ),
-    );
-  }
-}
-
-class _TotalsChip extends StatelessWidget {
-  final double income;
-  final double expense;
-  final String currency;
-  final String locale;
-  const _TotalsChip({
-    required this.income,
-    required this.expense,
-    required this.currency,
-    required this.locale,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final f = fmt.currencyFormat(locale, currency);
-    final net = income + expense; // expense is negative
-    if (income == 0 && expense == 0) return const SizedBox.shrink();
-    final scheme = Theme.of(context).colorScheme;
-    const numStyle = TextStyle(fontSize: 11, fontWeight: FontWeight.w600);
-    final parts = <Widget>[];
-    if (income > 0) {
-      parts.add(PrivacyText(
-        '+${f.format(income)}',
-        style: numStyle.copyWith(color: Colors.green.shade700),
-      ));
-    }
-    if (expense < 0) {
-      if (parts.isNotEmpty) parts.add(const SizedBox(width: 6));
-      parts.add(PrivacyText(
-        f.format(expense),
-        style: numStyle.copyWith(color: Colors.red.shade700),
-      ));
-    }
-    // Consolidated net — always shown so each row carries a definitive total.
-    if (parts.isNotEmpty) parts.add(const SizedBox(width: 8));
-    parts.add(PrivacyText(
-      '= ${net >= 0 ? '+' : ''}${f.format(net)}',
-      style: numStyle.copyWith(
-        color: net >= 0 ? Colors.green.shade700 : Colors.red.shade700,
-        fontWeight: FontWeight.w700,
-        decoration: TextDecoration.underline,
-        decorationColor: scheme.onSurfaceVariant.withValues(alpha: 0.35),
-      ),
-    ));
-    return Row(mainAxisSize: MainAxisSize.min, children: parts);
-  }
-}
